+++ src/server/index.js
@@ -1,6 +1,6 @@
 import '../shared/utils/peerDependencies';
 import { createRouter, debugLastRequest, debugLastResponse } from './createRouter';
-import logger from './utils/logger';
+import logger from '../shared/utils/logger';
 import pure from '../shared/hoc/pure';
 import asymetricSsr from '../shared/hoc/asymetricSsr';
 import {