+++ src/shared/constants/index.js
@@ -1,10 +1,8 @@
 import URL from './url';
 
-const LOG_LEVEL = 'debug';
 const NOT_FOUND_URL = '/not-found';
 
 export {
   URL,
-  LOG_LEVEL,
   NOT_FOUND_URL,
 };