+++ src/shared/utils/logger.js
@@ -1,0 +1,7 @@
+/* eslint-disable no-undef, import/no-extraneous-dependencies, import/no-unresolved, import/extensions, max-len */
+import pino from 'pino';
+/* eslint-enable */
+
+const logger = pino();
+
+export default logger;