+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * react-helmet: 5.0.0 API
 * `sitemap.xml` helper has been removed in favor of [sitemap.js](https://github.com/ekalinin/sitemap.js)
 * `humans.txt` has been added as available static routes
+* Replace logatim & winston with pino for universal logging